+++ packages/components/organisms/f-footer/CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## v8.9.0
+
+_June 24, 2024_
+
+### Added
+
+- New prop `globalTrackingContexts` which are passed through to f-trak for analytics events.
+- Use `f-trak` v1+
+
 ## v8.8.0
 
 _May 21, 2024_